Finished functionality of switching screens from main menu to game screen. Also packaged things a bit more.
