+++ src/main/java/game/screens/MainMenuScreen.java
@@ -1,0 +1,93 @@
+package game.screens;
+
+import com.badlogic.gdx.Game;
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Screen;
+
+import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.Texture;
+import game.GdxGame;
+
+public class MainMenuScreen implements Screen {
+
+    GdxGame game;
+
+    Texture exitButtonActive;
+    Texture exitButtonInActive;
+    Texture playButtonActive;
+    Texture playButtonInActive;
+
+    private static final int EXIT_BUTTON_WIDTH = 300;
+    private static final int EXIT_BUTTON_HEIGHT = 150;
+    private static final int PLAY_BUTTON_WIDTH = 300;
+    private static final int PLAY_BUTTON_HEIGHT = 150;
+    private static final int EXIT_BUTTON_X = GdxGame.WIDTH / 2 - EXIT_BUTTON_WIDTH / 2;
+    private static final int EXIT_BUTTON_Y = 100;
+    private static final int PLAY_BUTTON_X = GdxGame.WIDTH / 2 - PLAY_BUTTON_WIDTH / 2;
+    private static final int PLAY_BUTTON_Y = 300;
+
+
+    public MainMenuScreen(GdxGame game) {
+        this.game = game;
+        playButtonActive = new Texture("tiles/grass.png");
+        playButtonInActive = new Texture("tiles/path.png");
+        exitButtonActive = new Texture("tiles/grass.png");
+        exitButtonInActive = new Texture("tiles/wall.png");
+    }
+
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+        Gdx.gl.glClearColor(1, 0, 0, 1);
+        Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
+
+        game.batch.begin();
+
+        if(Gdx.input.getX() < EXIT_BUTTON_X + EXIT_BUTTON_WIDTH && Gdx.input.getX() > EXIT_BUTTON_X && GdxGame.HEIGHT - Gdx.input.getY() < EXIT_BUTTON_Y + EXIT_BUTTON_HEIGHT && GdxGame.HEIGHT - Gdx.input.getY() > EXIT_BUTTON_Y){
+            game.batch.draw(exitButtonActive, EXIT_BUTTON_X, EXIT_BUTTON_Y, EXIT_BUTTON_WIDTH, EXIT_BUTTON_HEIGHT);
+            if (Gdx.input.isTouched()){
+                Gdx.app.exit();
+            }
+        }else{
+            game.batch.draw(exitButtonInActive, EXIT_BUTTON_X, EXIT_BUTTON_Y, EXIT_BUTTON_WIDTH, EXIT_BUTTON_HEIGHT);
+        }
+
+        if(Gdx.input.getX() < PLAY_BUTTON_X + PLAY_BUTTON_WIDTH && Gdx.input.getX() > PLAY_BUTTON_X && GdxGame.HEIGHT - Gdx.input.getY() < PLAY_BUTTON_Y + PLAY_BUTTON_HEIGHT && GdxGame.HEIGHT - Gdx.input.getY() > PLAY_BUTTON_Y){
+            game.batch.draw(playButtonActive, PLAY_BUTTON_X, PLAY_BUTTON_Y, PLAY_BUTTON_WIDTH, PLAY_BUTTON_HEIGHT);
+            if(Gdx.input.isTouched()){
+                this.dispose();
+                game.setScreen(new GameScreen(game));
+            }
+        }else{
+            game.batch.draw(playButtonInActive, PLAY_BUTTON_X, PLAY_BUTTON_Y, PLAY_BUTTON_WIDTH, PLAY_BUTTON_HEIGHT);
+        }
+
+
+        game.batch.end();
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+    }
+
+    @Override
+    public void resume() {
+    }
+
+    @Override
+    public void hide() {
+    }
+
+    @Override
+    public void dispose() {
+    }
+}
